refactor: sync file name to addresses
