+++ src/pages/point-shop/products/[pointProductId]/enroll-address.tsx
@@ -4,7 +4,7 @@ import Input from '@/components/common/form/input'
 import InputLabel from '../../../../components/common/form/input-label'
 import BottomNavigation from '../../../../components/common/bottom-navigation'
 import { useNavigate, useSearchParams } from 'react-router-dom'
-import { addressApi } from '@/api/address'
+import { addressApi } from '@/api/addresses'
 import { ClientAddress } from '@/types/addresses'
 import PageLayOut from '@/components/common/page-layout'
 import PageTitle from '@/components/common/page-title'
